+++ CHANGELOG.md
@@ -8,6 +8,9 @@ x.y.z / 2014-06-dd
 * added coveralls
 * upgraded `bs58` from `0.3.x` to `^1.0.0`
 * removed `crypto-hashing` dep
+* `coinstring()` renamed to `cs.encode()`
+
+
 
 0.2.0 / 2014-03-10
 ------------------